tests/submission/Microsoft/forms/formenctype.html

<!DOCTYPE html>
<html><head>
        <title>formenctype on button element</title>
        <meta content="text/html; charset=UTF-8" http-equiv="content-type">
        <meta content="formenctype support on elements" name="description">
    <link href="http://www.w3.org/TR/html5/attributes-common-to-form-controls.html#dom-fs-formenctype" rel="help">
</head>
    <body>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>

        <div id="log"></div>
        <button formenctype="text/plain" style="display: none" type="submit">Submit</button>
        <input formenctype="text/plain" style="display: none" type="submit" value="submit">
        <input style="display: none" type="submit" value="submit">
        
        <script type="text/javascript">
        
            test(function() {assert_equals(document.getElementsByTagName("button")[0].formEnctype, "text/plain")}, "formEnctype support on button");
            test(function() {assert_equals(document.getElementsByTagName("input")[0].formEnctype, "text/plain")}, "formEnctype support on input");
        
            var testElem = document.getElementsByTagName("input")[0];
            testElem.formEnctype = "multipart/form-data";

            test(function() {assert_equals(document.getElementsByTagName("input")[0].formEnctype, "multipart/form-data")}, "formEnctype on input reflects correct value after being updated by the DOM");
            test(function() {assert_equals(document.getElementsByTagName("input")[0].getAttribute("formenctype"), "multipart/form-data")}, "formEnctype value is correct using getAttribute");
                
            test(function() {assert_equals(document.getElementsByTagName("input")[1].formEnctype, "application/x-www-form-urlencoded")}, "formEnctype defaults to application/x-www-form-urlencoded");
        </script>
    
</body></html>
